Add option to configure ignored repos
**Why**:
Since the bot is currently configured to post all open PRs across the 18F GitHub org, we want to reduce the noise by being able to ignore certain repos, namely ones that are active and presumably don't need outside help for reviews.

**How**:
Add support for a new entry in the YAML config file.

Based on the original code, this would have meant adding another assignment to `Seal#team_params(team)`, and increasing the number of parameters passed in to a `GithubFetcher.new`.

Rather than going down this path for every new configuration we want to add, I simplified the method so that it passes in the team hash directly to `GithubFetcher`. I also removed the option to use ENV-based configuration to keep things simple and consistent, and added a check to make sure a YAML config file is present before the bot can run.