Add language-based teams
**Why**:
To be able to break down the list of open PRs into separate buckets based on the language of the PR. That way, all Ruby PRs will go to the #ruby-pull-requests Slack channel, all Python PRs will go to #python-pull-requests, and so on.

**How**:
- Move the global configuration (title, labels, repos to exclude) to a separate `global.yml` file, and create language-based teams in the `18f.yml` file.
- Update the GitHub query to filter by the team's language

I tested this and it works great. Check out the slack channels listed in `18f.yml` to see what the bot posted.